Rename view diff modules
