+++ perun/view_diff/__init__.py
@@ -1,6 +1,7 @@
 """
 Base package for difference of profiles
 """
+
 from __future__ import annotations
 
 from typing import Callable, Any
@@ -11,15 +12,15 @@ def lazy_get_cli_commands() -> list[Callable[..., Any]]:
     Lazily imports CLI commands
     """
     import perun.view_diff.flamegraph.run as flamegraph_run
-    import perun.view_diff.report.run as report_run
+    import perun.view_diff.datatables.run as datatables_run
     import perun.view_diff.sankey.run as sankey_run
-    import perun.view_diff.sankey_incremental.run as sankey_incremental_run
-    import perun.view_diff.table.run as table_run
+    import perun.view_diff.report.run as report_run
+    import perun.view_diff.short.run as short_run
 
     return [
         flamegraph_run.flamegraph,
-        table_run.table,
-        report_run.report,
+        short_run.short,
+        datatables_run.datatables,
         sankey_run.sankey,
-        sankey_incremental_run.sankey_incr,
+        report_run.report,
     ]
