+++ perun/view_diff/datatables/__init__.py
@@ -1,0 +1,3 @@
+"""
+Base package for report difference of profiles
+"""
