+++ perun/view_diff/datatables/meson.build
@@ -1,0 +1,11 @@
+perun_view_diff_datatables_dir = perun_view_diff_dir / 'datatables'
+
+perun_view_diff_datatables_files = files(
+    '__init__.py',
+    'run.py',
+)
+
+py3.install_sources(
+    perun_view_diff_datatables_files,
+    subdir: perun_view_diff_datatables_dir
+)
