--- perun/view_diff/sankey_incremental/__init__.py
@@ -1,3 +1,0 @@
-"""
-Base package for sankey difference of profiles
-"""
