--- perun/view_diff/sankey_incremental/meson.build
@@ -1,11 +1,0 @@
-perun_view_diff_sankey_incremental_dir = perun_view_diff_dir / 'sankey_incremental'
-
-perun_view_diff_sankey_incremental_files = files(
-    '__init__.py',
-    'run.py',
-)
-
-py3.install_sources(
-    perun_view_diff_sankey_incremental_files,
-    subdir: perun_view_diff_sankey_incremental_dir
-)
