+++ perun/view_diff/short/meson.build
@@ -1,0 +1,11 @@
+perun_view_diff_short_dir = perun_view_diff_dir / 'short'
+
+perun_view_diff_short_files = files(
+    '__init__.py',
+    'run.py',
+)
+
+py3.install_sources(
+    perun_view_diff_short_files,
+    subdir: perun_view_diff_short_dir
+)
