--- perun/view_diff/table/meson.build
@@ -1,11 +1,0 @@
-perun_view_diff_table_dir = perun_view_diff_dir / 'table'
-
-perun_view_diff_table_files = files(
-    '__init__.py',
-    'run.py',
-)
-
-py3.install_sources(
-    perun_view_diff_table_files,
-    subdir: perun_view_diff_table_dir
-)
