+++ tests/conftest.py
@@ -20,7 +20,7 @@ from perun.logic import commands, pcs, store
 from perun.utils import decorators, log, metrics, streams
 from perun.utils.common import common_kit
 import perun.testing.utils as test_utils
-import perun.view_diff.sankey_incremental.run as sankey_incremental
+import perun.view_diff.report.run as report
 
 
 @pytest.fixture(scope="session", autouse=True)
@@ -529,6 +529,6 @@ def setup():
     log.SUPPRESS_PAGING = True
     # We disable the metrics by default, since they might slow down tests
     metrics.Metrics.enabled = False
-    sankey_incremental.Stats.KnownStatsSet.clear()
-    sankey_incremental.Stats.SortedStats = []
+    report.Stats.KnownStatsSet.clear()
+    report.Stats.SortedStats = []
     yield
